requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pyyaml==6.0.1
uritemplate==4.1.1
